+++ generators/generate_category.py
@@ -3,13 +3,6 @@ import sys, os
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 
 from dolibarr_api import *
-from generators.utils.fill_data import *
-from generators.utils.get_data import *
-from generators.utils.put_data import *
-from generators.utils.utils import *
-
-
-
 
 def generate_category(type, testing):
     hexColor= fake.hex_color()
