+++ generators/generate_customer.py
@@ -1,15 +1,11 @@
-from faker import Faker
+
 import random
 import requests
 import sys, os
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 
 from dolibarr_api import *
-from generators.utils.fill_data import *
-from generators.utils.get_data import *
-from generators.utils.put_data import *
-from generators.utils.utils import *
-
+from utils.get_random_address import *
 
 def generate_customer(dateCreate, retDataCategContact, retDataCategCustomer, enabledModule, testing):
     # on boucle sur les lignes
