+++ generators/generate_intervention.py
@@ -1,14 +1,9 @@
-from faker import Faker
 import random
 import requests
 import sys, os
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 
 from dolibarr_api import *
-from generators.utils.fill_data import *
-from generators.utils.get_data import *
-from generators.utils.put_data import *
-from generators.utils.utils import *
 
 
 def generate_intervention(dateIntervention, retDataThirdParties, enabledModule, testing):
@@ -19,7 +14,7 @@ def generate_intervention(dateIntervention, retDataThirdParties, enabledModule, 
     if 'contrat' in enabledModule:
         retDataContract = fill_contracts(socid)
         fk_contract = get_random_contract(retDataContract)
-        fk_project = get_random_project_id(get_projects_of_contactID(socid))
+        fk_project = 1#get_random_project_id(get_projects_of_contactID(socid))
 
     data = {
         "socid": socid,
