+++ generators/generate_invoice.py
@@ -5,10 +5,7 @@ import sys, os
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 
 from dolibarr_api import *
-from generators.utils.fill_data import *
-from generators.utils.get_data import *
-from generators.utils.put_data import *
-from generators.utils.utils import *
+
 
 def generate_invoice(dateFact,retDataPayment, retDataBank, retDataProduct, retDataThirdParties, retDataUser, testing):
     url = urlBase + "invoices"
@@ -16,17 +13,17 @@ def generate_invoice(dateFact,retDataPayment, retDataBank, retDataProduct, retDa
 
     paye = random.choice([0, 1])
     socId = get_random_client(retDataThirdParties)
+    fk_project = 1#get_random_project_id(get_projects_of_contactID(socid))
+
     data = {
         "type": "0",
         "date" :dateFact.strftime('%Y-%m-%d'),
         "socid": socId,
+        "fk_project" : fk_project
     }
     r = requests.post(url, headers=headers, json=data)
     invoiceID = r.text
     urlInvoice = url + "/" + str(invoiceID)
-
-     # on lie un projet du client à la commande
-    put_fk_project(socId, urlInvoice)
 
     # on ajoute les lignes
     urlLine = urlBase + "invoices/" + str(invoiceID) + "/lines"
