--- generators/utils/utils.py
@@ -1,3 +1,0 @@
-from datetime import datetime
-
-yearNow = datetime.now().year
